Fix function to return double instead of int

+++ configuration/pih/liquibase/sql/create_functions.sql
@@ -3670,12 +3670,12 @@ END
 DROP FUNCTION IF EXISTS obs_from_group_id_value_numeric_from_temp;
 #
 CREATE FUNCTION obs_from_group_id_value_numeric_from_temp(_obsGroupId int(11), _source varchar(50), _term varchar(255))
-    RETURNS int
+    RETURNS double
     DETERMINISTIC
 
 BEGIN
 
-    DECLARE ret int;
+    DECLARE ret double;
 
     select      value_numeric into ret
     from        temp_obs o
